requirements.txt
requests>=2.25.1
Pillow>=8.0.0
numpy>=1.19.5
opencv-python>=4.5.1
scikit-learn>=0.24.1
matplotlib>=3.3.3

# For PyTorch (choose based on your system)
torch>=1.9.0
torchvision>=0.10.0
